requirements.txt
django==2.2.9
SQLAlchemy==1.4.9
psycopg2==2.9.1
django-simple-history==2.8.0
djangorestframework==3.11.0
